add create account hook
